+++ packages/muelu/src/Graph/MatrixTransformation/MueLu_CoalesceDropFactory_kokkos_def.hpp
@@ -59,7 +59,7 @@
 #include "MueLu_LWGraph_kokkos.hpp"
 #include "MueLu_MasterList.hpp"
 #include "MueLu_Monitor.hpp"
-#include "MueLu_Utilities_kokkos.hpp"
+#include "MueLu_Utilities.hpp"
 
 namespace MueLu {
 
@@ -549,7 +549,7 @@ namespace MueLu {
       // Scalar problem without dropping
 
       // Detect and record rows that correspond to Dirichlet boundary conditions
-      boundaryNodes = Utilities_kokkos::DetectDirichletRows(*A, dirichletThreshold);
+      boundaryNodes = Utilities::DetectDirichletRows_kokkos(*A, dirichletThreshold);
 
       // Trivial LWGraph construction
       graph = rcp(new LWGraph_kokkos(A->getCrsGraph()->getLocalGraphDevice(), A->getRowMap(), A->getColMap(), "graph of A"));
@@ -619,7 +619,7 @@ namespace MueLu {
           {
             kokkosMatrix = local_matrix_type();
             SubFactoryMonitor m2(*this, "Ghosted diag construction", currentLevel);
-            ghostedDiag     = Utilities_kokkos::GetMatrixOverlappedDiagonal(*A);
+            ghostedDiag     = Utilities::GetMatrixOverlappedDiagonal(*A);
             kokkosMatrix=A->getLocalMatrixDevice();
           }
 
@@ -866,7 +866,7 @@ namespace MueLu {
       Kokkos::parallel_scan("MueLu:CoalesceDropF:Build:scalar_filter:stage1_scan", range_type(0,numNodes+1), scanFunctor);
 
       // Detect and record dof rows that correspond to Dirichlet boundary conditions
-      boundary_nodes_type singleEntryRows = Utilities_kokkos::DetectDirichletRows(*A, dirichletThreshold);
+      boundary_nodes_type singleEntryRows = Utilities::DetectDirichletRows_kokkos(*A, dirichletThreshold);
 
       typename entries_type::non_const_type dofcols("dofcols", numDofCols/*dofNnz(numNodes)*/); // why does dofNnz(numNodes) work? should be a parallel reduce, i guess
