+++ packages/muelu/src/Operators/MueLu_Maxwell_Utils_decl.hpp
@@ -99,6 +99,17 @@ namespace MueLu {
                                            Teuchos::ArrayRCP<bool> & BCdomain,
                                            bool & allEdgesBoundary,
                                            bool & allNodesBoundary);
+
+    static void detectBoundaryConditionsSM(RCP<Matrix> & SM_Matrix,
+                                           RCP<Matrix> & D0_Matrix,
+                                           magnitudeType rowSumTol,
+                                           Kokkos::View<bool*, typename Node::device_type> & BCrowsKokkos,
+                                           Kokkos::View<bool*, typename Node::device_type> & BCcolsKokkos,
+                                           Kokkos::View<bool*, typename Node::device_type> & BCdomainKokkos,
+                                           int & BCedges,
+                                           int & BCnodes,
+                                           bool & allEdgesBoundary,
+                                           bool & allNodesBoundary);
         
     //! Remove explicit zeros
     static void removeExplicitZeros(Teuchos::ParameterList &parameterList,
