+++ packages/muelu/src/Operators/MueLu_RefMaxwell_decl.hpp
@@ -399,9 +399,8 @@ namespace MueLu {
     Teuchos::RCP<const Map> D0origDomainMap_;
     Teuchos::RCP<const Import> D0origImporter_;
     //! Vectors for BCs
-    Kokkos::View<bool*, typename Node::device_type> BCrowsKokkos_, BCcolsKokkos_, BCdomainKokkos_;
+    Kokkos::View<bool*, typename Node::device_type> BCrows_, BCcols_, BCdomain_;
     int BCedges_, BCnodes_;
-    Teuchos::ArrayRCP<bool> BCrows_, BCcols_, BCdomain_;
     //! Nullspace
     Teuchos::RCP<MultiVector> Nullspace_;
     //! Coordinates
