+++ packages/muelu/src/Utils/MueLu_Utilities_kokkos_decl.hpp
@@ -142,84 +142,6 @@ namespace MueLu {
     */
     static RCP<Vector> GetMatrixOverlappedDiagonal(const Matrix& A); // FIXME
 
-    static void MyOldScaleMatrix(Matrix& Op, const Teuchos::ArrayRCP<const SC>& scalingVector, bool doInverse = true,
-                                 bool doFillComplete = true, bool doOptimizeStorage = true); // FIXME
-
-    static void MyOldScaleMatrix_Tpetra(Matrix& Op, const Teuchos::ArrayRCP<SC>& scalingVector,
-                                        bool doFillComplete, bool doOptimizeStorage); // FIXME
-
-    static void MyOldScaleMatrix_Epetra(Matrix& Op, const Teuchos::ArrayRCP<SC>& scalingVector,
-                                            bool doFillComplete, bool doOptimizeStorage); // FIXME
-
-    // static RCP<Teuchos::FancyOStream> MakeFancy(std::ostream& os)       { return Utilities::MakeFancy(os); }
-
-    /*! @brief Detect Dirichlet rows
-
-        @param[in] A matrix
-        @param[in] tol If a row entry's magnitude is less than or equal to this tolerance, the entry is treated as zero.
-
-        @return boolean array.  The ith entry is true iff row i is a Dirichlet row.
-    */
-    static Kokkos::View<bool*, typename NO::device_type> DetectDirichletRows(const Matrix& A, const Magnitude& tol = Teuchos::ScalarTraits<typename Teuchos::ScalarTraits<SC>::magnitudeType>::zero(), const bool count_twos_as_dirichlet=false);
-
-
-    /*! @brief Find non-zero values in an ArrayRCP
-      Compares the value to 2 * machine epsilon
-
-      @param[in]  vals - ArrayRCP<const Scalar> of values to be tested
-      @param[out] nonzeros - ArrayRCP<bool> of true/false values for whether each entry in vals is nonzero
-    */
-    static void FindNonZeros(const typename Xpetra::MultiVector<Scalar,LocalOrdinal,GlobalOrdinal,Node>::dual_view_type::t_dev_const_um vals,
-                             Kokkos::View<bool*, typename Node::device_type> nonzeros);
-
-
-    /*! @brief Detect Dirichlet columns based on Dirichlet rows
-
-        The routine finds all column indices that are in Dirichlet rows, where Dirichlet rows are described by dirichletRows,
-        as returned by DetectDirichletRows.
-
-        @param[in] A matrix
-        @param[in] dirichletRows array of Dirichlet rows.
-
-        @return boolean array.  The ith entry is true iff row i is a Dirichlet column.
-    */
-    static Kokkos::View<bool*, typename NO::device_type> DetectDirichletCols(const Matrix& A, const Kokkos::View<const bool*, typename NO::device_type>& dirichletRows);
-
-
-    /*! @brief Detects Dirichlet columns & domains from a list of Dirichlet rows
-
-      @param[in] A - Matrix on which to apply Dirichlet column detection
-      @param[in] dirichletRows - View<bool> of indicators as to which rows are Dirichlet
-      @param[out] dirichletCols - View<bool> of indicators as to which cols are Dirichlet
-      @param[out] dirichletDomain - View<bool> of indicators as to which domains are Dirichlet
-    */
-    static void DetectDirichletColsAndDomains(const Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node>& A,
-                                              const Kokkos::View<bool*, typename Node::device_type> & dirichletRows,
-                                              Kokkos::View<bool*, typename Node::device_type> dirichletCols,
-                                              Kokkos::View<bool*, typename Node::device_type> dirichletDomain);
-
-
-    static void ZeroDirichletRows(RCP<Matrix>& A, const Kokkos::View<const bool*, typename NO::device_type>& dirichletRows, SC replaceWith=Teuchos::ScalarTraits<SC>::zero());
-
-    static void ZeroDirichletRows(RCP<MultiVector>& X, const Kokkos::View<const bool*, typename NO::device_type>& dirichletRows, SC replaceWith=Teuchos::ScalarTraits<SC>::zero());
-
-    static void ZeroDirichletCols(RCP<Matrix>& A, const Kokkos::View<const bool*, typename NO::device_type>& dirichletCols, SC replaceWith=Teuchos::ScalarTraits<SC>::zero());
-
-    static void ApplyRowSumCriterion(const Matrix& A,
-                                     const typename Teuchos::ScalarTraits<Scalar>::magnitudeType rowSumTol,
-                                     Kokkos::View<bool*, typename NO::device_type> & dirichletRows);
-
-    /*! Perform a Cuthill-McKee (CM) or Reverse Cuthill-McKee (RCM) ordering of the local component of the matrix.
-      Kokkos-Kernels has an RCM implementation, so we reverse that here if we call CM.
-    */
-    static RCP<Xpetra::Vector<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node> > CuthillMcKee(const Matrix &Op);
-
-    /*! Perform a Reverse Cuthill-McKee (RCM) ordering of the local component of the matrix.
-     */
-    static RCP<Xpetra::Vector<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node> > ReverseCuthillMcKee(const Matrix &Op);
-
-    static void ApplyOAZToMatrixRows(RCP<Matrix>& A, const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows);
-
   }; // class Utils
 
 
@@ -256,352 +178,14 @@ namespace MueLu {
 
       return diag;
     }
+
     static RCP<Vector> GetMatrixDiagonalInverse(const Matrix& A, Magnitude tol = Teuchos::ScalarTraits<SC>::eps()*100, const bool doLumped=false);
 
-    // todo: move this to UtilitiesBase::kokkos
-    static Kokkos::View<bool*, typename Node::device_type> DetectDirichletRows(const Matrix& A, const Magnitude& tol = Teuchos::ScalarTraits<SC>::zero(), const bool count_twos_as_dirichlet=false);
-
-    static Kokkos::View<bool*, typename Node::device_type> DetectDirichletCols(const Matrix& A, const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows);
-
-    static void FindNonZeros(const typename Xpetra::MultiVector<Scalar,LocalOrdinal,GlobalOrdinal,Node>::dual_view_type::t_dev_const_um vals,
-                             Kokkos::View<bool*, typename Node::device_type> nonzeros);
-
-    static void DetectDirichletColsAndDomains(const Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node>& A,
-                                              const Kokkos::View<bool*, typename Node::device_type> & dirichletRows,
-                                              Kokkos::View<bool*, typename Node::device_type> dirichletCols,
-                                              Kokkos::View<bool*, typename Node::device_type> dirichletDomain);
-
-    static void ZeroDirichletRows(RCP<Matrix>& A, const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows, SC replaceWith=Teuchos::ScalarTraits<SC>::zero());
-
-    static void ZeroDirichletRows(RCP<MultiVector>& X, const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows, SC replaceWith=Teuchos::ScalarTraits<SC>::zero());
-
-    static void ZeroDirichletCols(RCP<Matrix>& A, const Kokkos::View<const bool*, typename Node::device_type>& dirichletCols, SC replaceWith=Teuchos::ScalarTraits<SC>::zero());
-
-    static void ApplyRowSumCriterion(const Matrix& A,
-                                     const typename Teuchos::ScalarTraits<Scalar>::magnitudeType rowSumTol,
-                                     Kokkos::View<bool*, typename NO::device_type> & dirichletRows);
-
-    static void MyOldScaleMatrix(Matrix& Op, const Teuchos::ArrayRCP<const SC>& scalingVector, bool doInverse = true,
-        bool doFillComplete = true, bool doOptimizeStorage = true) {
-      SC one = Teuchos::ScalarTraits<SC>::one();
-      Teuchos::ArrayRCP<SC> sv(scalingVector.size());
-      if (doInverse) {
-        for (int i = 0; i < scalingVector.size(); ++i)
-          sv[i] = one / scalingVector[i];
-      } else {
-        for (int i = 0; i < scalingVector.size(); ++i)
-          sv[i] = scalingVector[i];
-      }
-
-      switch (Op.getRowMap()->lib()) {
-      case Xpetra::UseTpetra:
-        MyOldScaleMatrix_Tpetra(Op, sv, doFillComplete, doOptimizeStorage);
-        break;
-
-      case Xpetra::UseEpetra:
-        MyOldScaleMatrix_Epetra(Op, sv, doFillComplete, doOptimizeStorage);
-        break;
-
-      default:
-        throw Exceptions::RuntimeError("Only Epetra and Tpetra matrices can be scaled.");
-#ifndef __NVCC__ //prevent nvcc warning
-        break;
-#endif
-      }
-    }
-
-    // TODO This is the <double,int,int> specialization
-    static void MyOldScaleMatrix_Tpetra(Matrix& Op, const Teuchos::ArrayRCP<Scalar>& scalingVector,
-        bool doFillComplete, bool doOptimizeStorage) {
-  #ifdef HAVE_MUELU_TPETRA
-  #ifdef HAVE_MUELU_TPETRA_INST_INT_INT
-      try {
-        Tpetra::CrsMatrix<SC,LO,GO,NO>& tpOp = Utilities::Op2NonConstTpetraCrs(Op);
-
-        const RCP<const Tpetra::Map<LO,GO,NO> > rowMap    = tpOp.getRowMap();
-        const RCP<const Tpetra::Map<LO,GO,NO> > domainMap = tpOp.getDomainMap();
-        const RCP<const Tpetra::Map<LO,GO,NO> > rangeMap  = tpOp.getRangeMap();
-
-        size_t maxRowSize = tpOp.getLocalMaxNumRowEntries();
-        if (maxRowSize == Teuchos::as<size_t>(-1)) // hasn't been determined yet
-          maxRowSize = 20;
-
-        std::vector<SC> scaledVals(maxRowSize);
-        if (tpOp.isFillComplete())
-          tpOp.resumeFill();
-
-        if (Op.isLocallyIndexed() == true) {
-	  typename Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>::local_inds_host_view_type cols;
-	  typename Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>::values_host_view_type vals;
-
-          for (size_t i = 0; i < rowMap->getLocalNumElements(); ++i) {
-            tpOp.getLocalRowView(i, cols, vals);
-            size_t nnz = tpOp.getNumEntriesInLocalRow(i);
-            if (nnz > maxRowSize) {
-              maxRowSize = nnz;
-              scaledVals.resize(maxRowSize);
-            }
-            for (size_t j = 0; j < nnz; ++j)
-              scaledVals[j] = vals[j]*scalingVector[i];
-
-            if (nnz > 0) {
-	      Teuchos::ArrayView<const LocalOrdinal> cols_view(cols.data(), nnz);
-              Teuchos::ArrayView<const SC> valview(&scaledVals[0], nnz);
-              tpOp.replaceLocalValues(i, cols_view, valview);
-            }
-          } //for (size_t i=0; ...
-
-        } else {
-          typename Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>::global_inds_host_view_type cols;
-          typename Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>::values_host_view_type vals;
-
-          for (size_t i = 0; i < rowMap->getLocalNumElements(); ++i) {
-            GO gid = rowMap->getGlobalElement(i);
-            tpOp.getGlobalRowView(gid, cols, vals);
-            size_t nnz = tpOp.getNumEntriesInGlobalRow(gid);
-            if (nnz > maxRowSize) {
-              maxRowSize = nnz;
-              scaledVals.resize(maxRowSize);
-            }
-            // FIXME FIXME FIXME FIXME FIXME FIXME
-            for (size_t j = 0; j < nnz; ++j)
-              scaledVals[j] = vals[j]*scalingVector[i]; //FIXME i or gid?
-
-	    if (nnz > 0) {
-	      Teuchos::ArrayView<const GlobalOrdinal> cols_view(cols.data(), nnz);
-              Teuchos::ArrayView<const SC> valview(&scaledVals[0], nnz);
-              tpOp.replaceGlobalValues(gid, cols_view, valview);
-            }
-          } //for (size_t i=0; ...
-        }
-
-        if (doFillComplete) {
-          if (domainMap == Teuchos::null || rangeMap == Teuchos::null)
-            throw Exceptions::RuntimeError("In Utilities::Scaling: cannot fillComplete because the domain and/or range map hasn't been defined");
-
-          RCP<Teuchos::ParameterList> params = rcp(new Teuchos::ParameterList());
-          params->set("Optimize Storage",    doOptimizeStorage);
-          params->set("No Nonlocal Changes", true);
-          Op.fillComplete(Op.getDomainMap(), Op.getRangeMap(), params);
-        }
-      } catch(...) {
-        throw Exceptions::RuntimeError("Only Tpetra::CrsMatrix types can be scaled (Err.1)");
-      }
-#else
-      throw Exceptions::RuntimeError("Matrix scaling is not possible because Tpetra has not been compiled with support for LO=GO=int.");
-#endif
-#else
-      throw Exceptions::RuntimeError("Matrix scaling is not possible because Tpetra has not been enabled.");
-#endif
-    }
-
-    static void MyOldScaleMatrix_Epetra(Matrix& Op, const Teuchos::ArrayRCP<Scalar>& scalingVector, bool doFillComplete, bool doOptimizeStorage) {
-#ifdef HAVE_MUELU_EPETRA
-      try {
-        //const Epetra_CrsMatrix& epOp = Utilities<double,int,int>::Op2NonConstEpetraCrs(Op);
-        const Epetra_CrsMatrix& epOp = Utilities::Op2NonConstEpetraCrs(Op);
-
-        Epetra_Map const &rowMap = epOp.RowMap();
-        int nnz;
-        double *vals;
-        int *cols;
-
-        for (int i = 0; i < rowMap.NumMyElements(); ++i) {
-          epOp.ExtractMyRowView(i, nnz, vals, cols);
-          for (int j = 0; j < nnz; ++j)
-            vals[j] *= scalingVector[i];
-        }
-
-      } catch (...){
-        throw Exceptions::RuntimeError("Only Epetra_CrsMatrix types can be scaled");
-      }
-#else
-      throw Exceptions::RuntimeError("Matrix scaling is not possible because Epetra has not been enabled.");
-#endif // HAVE_MUELU_EPETRA
-    }
-
-    /*! @brief Transpose a Xpetra::Matrix
-
-      Note: Currently, an error is thrown if the matrix isn't a Tpetra::CrsMatrix or Epetra_CrsMatrix.
-      In principle, however, we could allow any Epetra_RowMatrix because the Epetra transposer does.
-     */
-    static RCP<Matrix> Transpose(Matrix& Op, bool /* optimizeTranspose */ = false, const std::string & label = std::string(),const Teuchos::RCP<Teuchos::ParameterList> &params=Teuchos::null) {
-      switch (Op.getRowMap()->lib()) {
-      case Xpetra::UseTpetra:
-      {
-#ifdef HAVE_MUELU_TPETRA
-#ifdef HAVE_MUELU_TPETRA_INST_INT_INT
-        try {
-          const Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node>& tpetraOp = Utilities::Op2TpetraCrs(Op);
-
-          // Compute the transpose A of the Tpetra matrix tpetraOp.
-          RCP<Tpetra::CrsMatrix<Scalar, LocalOrdinal, GlobalOrdinal, Node> > A;
-          Tpetra::RowMatrixTransposer<SC,LO,GO,NO> transposer(rcpFromRef(tpetraOp),label);
-          {
-            using Teuchos::ParameterList;
-            using Teuchos::rcp;
-            RCP<ParameterList> transposeParams = params.is_null () ?
-              rcp (new ParameterList) :
-              rcp (new ParameterList (*params));
-            transposeParams->set ("sort", false);
-            A = transposer.createTranspose(transposeParams);
-          }
-
-          RCP<Xpetra::TpetraCrsMatrix<SC,LO,GO,NO> > AA   = rcp(new Xpetra::TpetraCrsMatrix<SC,LO,GO,NO>(A));
-          RCP<CrsMatrix>                             AAA  = rcp_implicit_cast<CrsMatrix>(AA);
-          RCP<CrsMatrixWrap>                         AAAA = rcp(new CrsMatrixWrap(AAA));
-
-          return AAAA;
-        }
-        catch (std::exception& e) {
-          std::cout << "threw exception '" << e.what() << "'" << std::endl;
-          throw Exceptions::RuntimeError("Utilities::Transpose failed, perhaps because matrix is not a Crs matrix");
-        }
-#else
-        throw Exceptions::RuntimeError("Utilities::Transpose: Tpetra is not compiled with LO=GO=int. Add TPETRA_INST_INT_INT:BOOL=ON to your configuration!");
-#endif
-#else
-        throw Exceptions::RuntimeError("Utilities::Transpose: Tpetra is not compiled!");
-#endif
-#ifndef __NVCC__ //prevent nvcc warning
-        break;
-#endif
-      }
-      case Xpetra::UseEpetra:
-      {
-#if defined(HAVE_MUELU_EPETRA) && defined(HAVE_MUELU_EPETRAEXT)
-        Teuchos::TimeMonitor tm(*Teuchos::TimeMonitor::getNewTimer("ZZ Entire Transpose"));
-        // Epetra case
-        Epetra_CrsMatrix& epetraOp = Utilities::Op2NonConstEpetraCrs(Op);
-        EpetraExt::RowMatrix_Transpose transposer;
-        Epetra_CrsMatrix * A = dynamic_cast<Epetra_CrsMatrix*>(&transposer(epetraOp));
-        transposer.ReleaseTranspose(); // So we can keep A in Muelu...
-
-        RCP<Epetra_CrsMatrix> rcpA(A);
-        RCP<Xpetra::EpetraCrsMatrixT<GO,NO> >  AA   = rcp(new Xpetra::EpetraCrsMatrixT<GO,NO> (rcpA));
-        RCP<CrsMatrix>                         AAA  = rcp_implicit_cast<CrsMatrix>(AA);
-        RCP<CrsMatrixWrap>                     AAAA = rcp( new CrsMatrixWrap(AAA));
-        AAAA->fillComplete(Op.getRangeMap(), Op.getDomainMap());
-
-        return AAAA;
-#else
-        throw Exceptions::RuntimeError("Epetra (Err. 2)");
-#endif
-#ifndef __NVCC__ //prevent nvcc warning
-        break;
-#endif
-      }
-      default:
-        throw Exceptions::RuntimeError("Only Epetra and Tpetra matrices can be transposed.");
-#ifndef __NVCC__ //prevent nvcc warning
-        break;
-#endif
-      }
-
-#ifndef __NVCC__ //prevent nvcc warning
-      return Teuchos::null;
-#endif
-    }
-
-    /*! @brief Extract coordinates from parameter list and return them in a Xpetra::MultiVector
-    */
-    static RCP<Xpetra::MultiVector<typename Teuchos::ScalarTraits<Scalar>::magnitudeType,LO,GO,NO> > ExtractCoordinatesFromParameterList(ParameterList& paramList) {
-      RCP<Xpetra::MultiVector<typename Teuchos::ScalarTraits<Scalar>::magnitudeType,LO,GO,NO> > coordinates = Teuchos::null;
-
-      // check whether coordinates are contained in parameter list
-      if(paramList.isParameter ("Coordinates") == false)
-        return coordinates;
-
-#if defined(HAVE_MUELU_TPETRA)
-#if ( defined(EPETRA_HAVE_OMP) && defined(HAVE_TPETRA_INST_OPENMP) && defined(HAVE_TPETRA_INST_INT_INT)) || \
-      (!defined(EPETRA_HAVE_OMP) && defined(HAVE_TPETRA_INST_SERIAL) && defined(HAVE_TPETRA_INST_INT_INT))
-
-      // define Tpetra::MultiVector type with Scalar=float only if
-      // * ETI is turned off, since then the compiler will instantiate it automatically OR
-      // * Tpetra is instantiated on Scalar=float
-#if !defined(HAVE_TPETRA_EXPLICIT_INSTANTIATION) || defined(HAVE_TPETRA_INST_FLOAT)
-      typedef Tpetra::MultiVector<float, LO,GO,NO> tfMV;
-      RCP<tfMV> floatCoords = Teuchos::null;
-#endif
-
-      // define Tpetra::MultiVector type with Scalar=double only if
-      // * ETI is turned off, since then the compiler will instantiate it automatically OR
-      // * Tpetra is instantiated on Scalar=double
-      typedef Tpetra::MultiVector<SC,LO,GO,NO> tdMV;
-      RCP<tdMV> doubleCoords = Teuchos::null;
-      if (paramList.isType<RCP<tdMV> >("Coordinates")) {
-        // Coordinates are stored as a double vector
-        doubleCoords = paramList.get<RCP<tdMV> >("Coordinates");
-        paramList.remove("Coordinates");
-      }
-#if !defined(HAVE_TPETRA_EXPLICIT_INSTANTIATION) || defined(HAVE_TPETRA_INST_FLOAT)
-      else if (paramList.isType<RCP<tfMV> >("Coordinates")) {
-        // check if coordinates are stored as a float vector
-        floatCoords = paramList.get<RCP<tfMV> >("Coordinates");
-        paramList.remove("Coordinates");
-        doubleCoords = rcp(new tdMV(floatCoords->getMap(), floatCoords->getNumVectors()));
-        deep_copy(*doubleCoords, *floatCoords);
-      }
-#endif
-      // We have the coordinates in a Tpetra double vector
-      if(doubleCoords != Teuchos::null) {
-        coordinates = Teuchos::rcp(new Xpetra::TpetraMultiVector<SC,LO,GO,NO>(doubleCoords));
-        TEUCHOS_TEST_FOR_EXCEPT(doubleCoords->getNumVectors() != coordinates->getNumVectors());
-      }
-#endif // Tpetra instantiated on GO=int and EpetraNode
-#endif // endif HAVE_TPETRA
-
-#if defined(HAVE_MUELU_EPETRA)
-      RCP<Epetra_MultiVector> doubleEpCoords;
-      if (paramList.isType<RCP<Epetra_MultiVector> >("Coordinates")) {
-        doubleEpCoords = paramList.get<RCP<Epetra_MultiVector> >("Coordinates");
-        paramList.remove("Coordinates");
-        RCP<Xpetra::EpetraMultiVectorT<GO,NO> > epCoordinates = Teuchos::rcp(new Xpetra::EpetraMultiVectorT<GO,NO>(doubleEpCoords));
-        coordinates = rcp_dynamic_cast<Xpetra::MultiVector<typename Teuchos::ScalarTraits<Scalar>::magnitudeType,LO,GO,NO> >(epCoordinates);
-        TEUCHOS_TEST_FOR_EXCEPT(doubleEpCoords->NumVectors() != Teuchos::as<int>(coordinates->getNumVectors()));
-      }
-#endif
-      TEUCHOS_TEST_FOR_EXCEPT(Teuchos::is_null(coordinates));
-      return coordinates;
-    }
-
-    /*! Perform a Cuthill-McKee (CM) or Reverse Cuthill-McKee (RCM) ordering of the local component of the matrix
-      Kokkos-Kernels has an RCM implementation, so we reverse that here if we call CM.
-     */
-    static RCP<Xpetra::Vector<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node> > ReverseCuthillMcKee(const Matrix &Op);
-
-    /*! Perform a Cuthill-McKee (CM) or Reverse Cuthill-McKee (RCM) ordering of the local component of the matrix
-      Kokkos-Kernels has an RCM implementation, so we reverse that here if we call CM.
-    */
-    static RCP<Xpetra::Vector<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node> > CuthillMcKee(const Matrix &Op);
-
-    static void ApplyOAZToMatrixRows(RCP<Matrix>& A, const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows);
 
   }; // class Utilities (specialization SC=double LO=GO=int)
 
 
 
-  // Useful Kokkos conversions
-  template < class View, unsigned AppendValue >
-  struct AppendTrait {
-    // static_assert(false, "Error: NOT a Kokkos::View");
-  };
-
-  template < class MT, unsigned T >
-  struct CombineMemoryTraits {
-    // empty
-  };
-
-  template < unsigned U, unsigned T>
-  struct CombineMemoryTraits<Kokkos::MemoryTraits<U>, T> {
-    typedef Kokkos::MemoryTraits<U|T> type;
-  };
-
-  template < class DataType, unsigned T, class... Pack >
-  struct AppendTrait< Kokkos::View< DataType, Pack... >, T> {
-    typedef Kokkos::View< DataType, Pack... > view_type;
-    using type = Kokkos::View< DataType, typename view_type::array_layout, typename view_type::device_type, typename CombineMemoryTraits<typename view_type::memory_traits,T>::type >;
-  };
 
 } //namespace MueLu
