+++ packages/muelu/src/Utils/MueLu_Utilities_kokkos_def.hpp
@@ -103,6 +103,7 @@
 #include <Xpetra_VectorFactory.hpp>
 
 #include <MueLu_Utilities_kokkos_decl.hpp>
+#include <MueLu_Utilities.hpp>
 
 #include <KokkosKernels_Handle.hpp>
 #include <KokkosGraph_RCM.hpp>
@@ -245,719 +246,8 @@ namespace MueLu {
     return diagonal;
   } //GetMatrixOverlappedDiagonal
 
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
-  MyOldScaleMatrix(Matrix& Op, const Teuchos::ArrayRCP<const SC>& scalingVector,
-                   bool doInverse, bool doFillComplete, bool doOptimizeStorage)
-  {
-    SC one = Teuchos::ScalarTraits<SC>::one();
-    Teuchos::ArrayRCP<SC> sv(scalingVector.size());
-    if (doInverse) {
-      for (int i = 0; i < scalingVector.size(); ++i)
-        sv[i] = one / scalingVector[i];
-    } else {
-      for (int i = 0; i < scalingVector.size(); ++i)
-        sv[i] = scalingVector[i];
-    }
 
-    switch (Op.getRowMap()->lib()) {
-      case Xpetra::UseTpetra:
-        MyOldScaleMatrix_Tpetra(Op, sv, doFillComplete, doOptimizeStorage);
-        break;
 
-      case Xpetra::UseEpetra:
-        // FIXME_KOKKOS
-        // Utils2_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::MyOldScaleMatrix_Epetra(Op, sv, doFillComplete, doOptimizeStorage);
-        throw std::runtime_error("FIXME");
-#ifndef __NVCC__ //prevent nvcc warning
-        break;
-#endif
-
-      default:
-        throw Exceptions::RuntimeError("Only Epetra and Tpetra matrices can be scaled.");
-#ifndef __NVCC__ //prevent nvcc warning
-        break;
-#endif
-    }
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::MyOldScaleMatrix_Epetra(Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node>& /* Op */, const Teuchos::ArrayRCP<Scalar>& /* scalingVector */, bool /* doFillComplete */, bool /* doOptimizeStorage */) {
-    throw Exceptions::RuntimeError("MyOldScaleMatrix_Epetra: Epetra needs SC=double and LO=GO=int.");
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::MyOldScaleMatrix_Tpetra(Matrix& Op, const Teuchos::ArrayRCP<SC>& scalingVector,
-                               bool doFillComplete,
-                               bool doOptimizeStorage)
-  {
-#ifdef HAVE_MUELU_TPETRA
-    try {
-      Tpetra::CrsMatrix<SC,LO,GO,NO>& tpOp = Op2NonConstTpetraCrs(Op);
-
-      const RCP<const Tpetra::Map<LO,GO,NO> > rowMap    = tpOp.getRowMap();
-      const RCP<const Tpetra::Map<LO,GO,NO> > domainMap = tpOp.getDomainMap();
-      const RCP<const Tpetra::Map<LO,GO,NO> > rangeMap  = tpOp.getRangeMap();
-
-      size_t maxRowSize = tpOp.getLocalMaxNumRowEntries();
-      if (maxRowSize == Teuchos::as<size_t>(-1)) // hasn't been determined yet
-        maxRowSize = 20;
-
-      if (tpOp.isFillComplete())
-        tpOp.resumeFill();
-
-      if (Op.isLocallyIndexed() == true) {
-	typename Tpetra::CrsMatrix<SC,LO,GO,NO>::local_inds_host_view_type cols;
-	typename Tpetra::CrsMatrix<SC,LO,GO,NO>::values_host_view_type vals;
-
-        for (size_t i = 0; i < rowMap->getLocalNumElements(); ++i) {
-          tpOp.getLocalRowView(i, cols, vals);
-          size_t nnz = tpOp.getNumEntriesInLocalRow(i);
-	  typename Tpetra::CrsMatrix<SC,LO,GO,NO>::nonconst_values_host_view_type scaledVals("scaledVals", nnz);
-          for (size_t j = 0; j < nnz; ++j)
-            scaledVals[j] = scalingVector[i]*vals[j];
-
-          if (nnz > 0) {
-            tpOp.replaceLocalValues(i, cols, scaledVals);
-          }
-        } //for (size_t i=0; ...
-
-      } else {
-	typename Tpetra::CrsMatrix<SC,LO,GO,NO>::global_inds_host_view_type cols;
-	typename Tpetra::CrsMatrix<SC,LO,GO,NO>::values_host_view_type vals;
-
-        for (size_t i = 0; i < rowMap->getLocalNumElements(); ++i) {
-          GO gid = rowMap->getGlobalElement(i);
-          tpOp.getGlobalRowView(gid, cols, vals);
-          size_t nnz = tpOp.getNumEntriesInGlobalRow(gid);
-	  typename Tpetra::CrsMatrix<SC,LO,GO,NO>::nonconst_values_host_view_type scaledVals("scaledVals", nnz);
-
-          // FIXME FIXME FIXME FIXME FIXME FIXME
-          for (size_t j = 0; j < nnz; ++j)
-            scaledVals[j] = scalingVector[i]*vals[j]; //FIXME i or gid?
-
-          if (nnz > 0) {
-            tpOp.replaceGlobalValues(gid, cols, scaledVals);
-          }
-        } //for (size_t i=0; ...
-      }
-
-      if (doFillComplete) {
-        if (domainMap == Teuchos::null || rangeMap == Teuchos::null)
-          throw Exceptions::RuntimeError("In Utils_kokkos::Scaling: cannot fillComplete because the domain and/or range map hasn't been defined");
-
-        RCP<Teuchos::ParameterList> params = rcp(new Teuchos::ParameterList());
-        params->set("Optimize Storage",    doOptimizeStorage);
-        params->set("No Nonlocal Changes", true);
-        Op.fillComplete(Op.getDomainMap(), Op.getRangeMap(), params);
-      }
-    } catch(...) {
-      throw Exceptions::RuntimeError("Only Tpetra::CrsMatrix types can be scaled (Err.1)");
-    }
-#else
-    throw Exceptions::RuntimeError("Matrix scaling is not possible because Tpetra has not been enabled.");
-#endif
-  } //MyOldScaleMatrix_Tpetra()
-
-
-  template <class SC, class LO, class GO, class NO>
-  Kokkos::View<bool*, typename NO::device_type>
-  DetectDirichletRows(const Xpetra::Matrix<SC,LO,GO,NO>& A,
-                      const typename Teuchos::ScalarTraits<SC>::magnitudeType& tol,
-                      const bool count_twos_as_dirichlet) {
-    using impl_scalar_type = typename Kokkos::ArithTraits<SC>::val_type;
-    using ATS        = Kokkos::ArithTraits<impl_scalar_type>;
-    using range_type = Kokkos::RangePolicy<LO, typename NO::execution_space>;
-    using helpers    = Xpetra::Helpers<SC,LO,GO,NO>;
-
-
-    if(helpers::isTpetraBlockCrs(A)) {
-#ifdef HAVE_MUELU_TPETRA
-      const Tpetra::BlockCrsMatrix<SC,LO,GO,NO> & Am = helpers::Op2TpetraBlockCrs(A);
-      auto b_graph      = Am.getCrsGraph().getLocalGraphDevice();
-      auto b_rowptr     = Am.getCrsGraph().getLocalRowPtrsDevice();
-      auto values       = Am.getValuesDevice();
-      LO   numBlockRows = Am.getLocalNumRows();
-      const LO stride   = Am.getBlockSize() * Am.getBlockSize();
-
-      Kokkos::View<bool*, typename NO::device_type> boundaryNodes(Kokkos::ViewAllocateWithoutInitializing("boundaryNodes"), numBlockRows);
-
-      if (count_twos_as_dirichlet)
-        throw Exceptions::RuntimeError("BlockCrs does not support counting twos as Dirichlet");
-
-      Kokkos::parallel_for("MueLu:Utils::DetectDirichletRowsBlockCrs", range_type(0,numBlockRows),
-                           KOKKOS_LAMBDA(const LO row) {
-                             auto rowView = b_graph.rowConst(row);
-                             auto length  = rowView.length;
-                             LO valstart  = b_rowptr[row] * stride;
-
-                             boundaryNodes(row) = true;
-                             decltype(length) colID =0;
-                             for (; colID < length; colID++) {
-                               if (rowView.colidx(colID) != row) {
-                                 LO current = valstart + colID*stride;
-                                 for(LO k=0; k<stride; k++) {
-                                   if (ATS::magnitude(values[current+ k]) > tol) {
-                                     boundaryNodes(row) = false;
-                                     break;
-                                   }
-                                 }
-                               }
-                               if(boundaryNodes(row) == false)
-                                 break;
-                             }
-                           });
-
-      return boundaryNodes;
-#else
-      throw Exceptions::RuntimeError("BlockCrs requires Tpetra");
-#endif
-    }
-    else {
-      auto localMatrix = A.getLocalMatrixDevice();
-      LO   numRows     = A.getLocalNumRows();
-      Kokkos::View<bool*, typename NO::device_type> boundaryNodes(Kokkos::ViewAllocateWithoutInitializing("boundaryNodes"), numRows);
-
-      if (count_twos_as_dirichlet)
-        Kokkos::parallel_for("MueLu:Utils::DetectDirichletRows_Twos_As_Dirichlet", range_type(0,numRows),
-                             KOKKOS_LAMBDA(const LO row) {
-                               auto rowView = localMatrix.row(row);
-                               auto length  = rowView.length;
-                               
-                               boundaryNodes(row) = true;
-                               if (length > 2) {
-                                 decltype(length) colID =0;
-                                 for ( ; colID < length; colID++)
-                                   if ((rowView.colidx(colID) != row) &&
-                                       (ATS::magnitude(rowView.value(colID)) > tol)) {
-                                     if (!boundaryNodes(row))
-                                       break;
-                                     boundaryNodes(row) = false;
-                                   }
-                                 if (colID == length)
-                                   boundaryNodes(row) = true;
-                               }
-                             });
-      else
-        Kokkos::parallel_for("MueLu:Utils::DetectDirichletRows", range_type(0,numRows),
-                             KOKKOS_LAMBDA(const LO row) {
-                               auto rowView = localMatrix.row(row);
-                               auto length  = rowView.length;
-                               
-                               boundaryNodes(row) = true;
-                               for (decltype(length) colID = 0; colID < length; colID++)
-                                 if ((rowView.colidx(colID) != row) &&
-                                     (ATS::magnitude(rowView.value(colID)) > tol)) {
-                                   boundaryNodes(row) = false;
-                                   break;
-                                 }
-                             });
-    return boundaryNodes;
-    }
-
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  Kokkos::View<bool*, typename Node::device_type>
-  Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
-  DetectDirichletRows(const Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node>& A, const typename Teuchos::ScalarTraits<Scalar>::magnitudeType& tol, const bool count_twos_as_dirichlet) {
-    return MueLu::DetectDirichletRows<Scalar,LocalOrdinal,GlobalOrdinal,Node>(A, tol, count_twos_as_dirichlet);
-  }
-
-  template <class Node>
-  Kokkos::View<bool*, typename Node::device_type>
-  Utilities_kokkos<double,int,int,Node>::
-  DetectDirichletRows(const Xpetra::Matrix<double,int,int,Node>& A, const typename Teuchos::ScalarTraits<double>::magnitudeType& tol, const bool count_twos_as_dirichlet) {
-    return MueLu::DetectDirichletRows<double,int,int,Node>(A, tol,count_twos_as_dirichlet);
-  }
-
-
-  template <class SC, class LO, class GO, class NO>
-  Kokkos::View<bool*, typename NO::device_type>
-  DetectDirichletCols(const Xpetra::Matrix<SC,LO,GO,NO>& A,
-                      const Kokkos::View<const bool*, typename NO::device_type>& dirichletRows) {
-    using ATS        = Kokkos::ArithTraits<SC>;
-    using impl_ATS = Kokkos::ArithTraits<typename ATS::val_type>;
-    using range_type = Kokkos::RangePolicy<LO, typename NO::execution_space>;
-
-    SC zero = ATS::zero();
-
-    auto localMatrix = A.getLocalMatrixDevice();
-    LO   numRows     = A.getLocalNumRows();
-
-    Teuchos::RCP<const Xpetra::Map<LO,GO,NO> > domMap = A.getDomainMap();
-    Teuchos::RCP<const Xpetra::Map<LO,GO,NO> > colMap = A.getColMap();
-    Teuchos::RCP<Xpetra::MultiVector<SC,LO,GO,NO> > myColsToZero = Xpetra::MultiVectorFactory<SC,LO,GO,NO>::Build(colMap,1);
-    myColsToZero->putScalar(zero);
-    auto myColsToZeroView = myColsToZero->getDeviceLocalView(Xpetra::Access::ReadWrite);
-    // Find all local column indices that are in Dirichlet rows, record in myColsToZero as 1.0
-    Kokkos::parallel_for("MueLu:Utils::DetectDirichletCols1", range_type(0,numRows),
-                         KOKKOS_LAMBDA(const LO row) {
-                           if (dirichletRows(row)) {
-                             auto rowView = localMatrix.row(row);
-                             auto length  = rowView.length;
-
-                             for (decltype(length) colID = 0; colID < length; colID++)
-                               myColsToZeroView(rowView.colidx(colID),0) = impl_ATS::one();
-                           }
-                         });
-
-    Teuchos::RCP<Xpetra::MultiVector<SC,LO,GO,NO> > globalColsToZero = Xpetra::MultiVectorFactory<SC,LO,GO,NO>::Build(domMap,1);
-    globalColsToZero->putScalar(zero);
-    Teuchos::RCP<Xpetra::Export<LO,GO,NO> > exporter = Xpetra::ExportFactory<LO,GO,NO>::Build(colMap,domMap);
-    // export to domain map
-    globalColsToZero->doExport(*myColsToZero,*exporter,Xpetra::ADD);
-    // import to column map
-    myColsToZero->doImport(*globalColsToZero,*exporter,Xpetra::INSERT);
-
-    auto myCols = myColsToZero->getDeviceLocalView(Xpetra::Access::ReadOnly);
-    size_t numColEntries = colMap->getLocalNumElements();
-    Kokkos::View<bool*, typename NO::device_type> dirichletCols(Kokkos::ViewAllocateWithoutInitializing("dirichletCols"), numColEntries);
-    const typename ATS::magnitudeType eps = 2.0*ATS::eps();
-
-    Kokkos::parallel_for("MueLu:Utils::DetectDirichletCols2", range_type(0,numColEntries),
-                         KOKKOS_LAMBDA (const size_t i) {
-                           dirichletCols(i) = impl_ATS::magnitude(myCols(i,0))>eps;
-                         });
-    return dirichletCols;
-  }
-
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  Kokkos::View<bool*, typename Node::device_type>
-  Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
-  DetectDirichletCols(const Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node>& A,
-                      const Kokkos::View<const bool*, typename NO::device_type>& dirichletRows) {
-    return MueLu::DetectDirichletCols<Scalar,LocalOrdinal,GlobalOrdinal,Node>(A, dirichletRows);
-  }
-
-  template <class Node>
-  Kokkos::View<bool*, typename Node::device_type>
-  Utilities_kokkos<double,int,int,Node>::
-  DetectDirichletCols(const Xpetra::Matrix<double,int,int,Node>& A,
-                      const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows) {
-    return MueLu::DetectDirichletCols<double,int,int,Node>(A, dirichletRows);
-  }
-
-
-  // Find Nonzeros in a device view
-  template<class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void FindNonZeros(const typename Xpetra::MultiVector<Scalar,LocalOrdinal,GlobalOrdinal,Node>::dual_view_type::t_dev_const_um vals,
-                    Kokkos::View<bool*, typename Node::device_type> nonzeros) {
-    using ATS        = Kokkos::ArithTraits<Scalar>;
-    using impl_ATS = Kokkos::ArithTraits<typename ATS::val_type>;
-    using range_type = Kokkos::RangePolicy<LocalOrdinal, typename Node::execution_space>;
-    TEUCHOS_ASSERT(vals.extent(0) == nonzeros.extent(0));
-    const typename ATS::magnitudeType eps = 2.0*impl_ATS::eps();
-
-    Kokkos::parallel_for("MueLu:Maxwell1::FindNonZeros", range_type(0,vals.extent(0)),
-                         KOKKOS_LAMBDA (const size_t i) {
-                           nonzeros(i) = (impl_ATS::magnitude(vals(i,0)) > eps);
-                         });
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void
-  Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
-  FindNonZeros(const typename Xpetra::MultiVector<Scalar,LocalOrdinal,GlobalOrdinal,Node>::dual_view_type::t_dev_const_um vals,
-                    Kokkos::View<bool*, typename Node::device_type> nonzeros) {
-    MueLu::FindNonZeros<Scalar,LocalOrdinal,GlobalOrdinal,Node>(vals,nonzeros);
-  }
-
-  template <class Node>
-  void
-  Utilities_kokkos<double,int,int,Node>::
-  FindNonZeros(const typename Xpetra::MultiVector<double,int,int,Node>::dual_view_type::t_dev_const_um vals,
-               Kokkos::View<bool*, typename Node::device_type> nonzeros) {
-    MueLu::FindNonZeros<double,int,int,Node>(vals,nonzeros);
-  }
-
-  // Detect Dirichlet cols and domains
-  template<class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void DetectDirichletColsAndDomains(const Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node>& A,
-                                     const Kokkos::View<bool*, typename Node::device_type> & dirichletRows,
-                                     Kokkos::View<bool*, typename Node::device_type> dirichletCols,
-                                     Kokkos::View<bool*, typename Node::device_type> dirichletDomain) {
-    using ATS        = Kokkos::ArithTraits<Scalar>;
-    using impl_ATS = Kokkos::ArithTraits<typename ATS::val_type>;
-    using range_type = Kokkos::RangePolicy<LocalOrdinal, typename Node::execution_space>;
-    RCP<const Xpetra::Map<LocalOrdinal,GlobalOrdinal,Node> > domMap = A.getDomainMap();
-    RCP<const Xpetra::Map<LocalOrdinal,GlobalOrdinal,Node> > rowMap = A.getRowMap();
-    RCP<const Xpetra::Map<LocalOrdinal,GlobalOrdinal,Node> > colMap = A.getColMap();
-    TEUCHOS_ASSERT(dirichletRows.extent(0) == rowMap->getLocalNumElements());
-    TEUCHOS_ASSERT(dirichletCols.extent(0) == colMap->getLocalNumElements());
-    TEUCHOS_ASSERT(dirichletDomain.extent(0) == domMap->getLocalNumElements());
-    RCP<Xpetra::Vector<Scalar,LocalOrdinal,GlobalOrdinal,Node> > myColsToZero = Xpetra::VectorFactory<Scalar,LocalOrdinal,GlobalOrdinal,Node>::Build(colMap, /*zeroOut=*/true);
-    // Find all local column indices that are in Dirichlet rows, record in myColsToZero as 1.0
-    auto myColsToZeroView = myColsToZero->getDeviceLocalView(Xpetra::Access::ReadWrite);
-    auto localMatrix = A.getLocalMatrixDevice();
-    Kokkos::parallel_for("MueLu:Maxwell1::DetectDirichletCols", range_type(0,rowMap->getLocalNumElements()),
-                         KOKKOS_LAMBDA(const LocalOrdinal row) {
-                           if (dirichletRows(row)) {
-                             auto rowView = localMatrix.row(row);
-                             auto length  = rowView.length;
-
-                             for (decltype(length) colID = 0; colID < length; colID++)
-                               myColsToZeroView(rowView.colidx(colID),0) = impl_ATS::one();
-                           }
-                         });
-
-    RCP<Xpetra::Vector<Scalar,LocalOrdinal,GlobalOrdinal,Node> > globalColsToZero;
-    RCP<const Xpetra::Import<LocalOrdinal,GlobalOrdinal,Node> > importer = A.getCrsGraph()->getImporter();
-    if (!importer.is_null()) {
-      globalColsToZero = Xpetra::VectorFactory<Scalar,LocalOrdinal,GlobalOrdinal,Node>::Build(domMap, /*zeroOut=*/true);
-      // export to domain map
-      globalColsToZero->doExport(*myColsToZero,*importer,Xpetra::ADD);
-      // import to column map
-      myColsToZero->doImport(*globalColsToZero,*importer,Xpetra::INSERT);
-    }
-    else
-      globalColsToZero = myColsToZero;
-    FindNonZeros<Scalar,LocalOrdinal,GlobalOrdinal,Node>(globalColsToZero->getDeviceLocalView(Xpetra::Access::ReadOnly),dirichletDomain);
-    FindNonZeros<Scalar,LocalOrdinal,GlobalOrdinal,Node>(myColsToZero->getDeviceLocalView(Xpetra::Access::ReadOnly),dirichletCols);
-  }
-
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void
-  Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
-  DetectDirichletColsAndDomains(const Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node>& A,
-                                const Kokkos::View<bool*, typename Node::device_type> & dirichletRows,
-                                Kokkos::View<bool*, typename Node::device_type> dirichletCols,
-                                Kokkos::View<bool*, typename Node::device_type> dirichletDomain) {
-    MueLu::DetectDirichletColsAndDomains<Scalar,LocalOrdinal,GlobalOrdinal,Node>(A,dirichletRows,dirichletCols,dirichletDomain);
-  }
-
-  template <class Node>
-  void
-  Utilities_kokkos<double,int,int,Node>::
-  DetectDirichletColsAndDomains(const Xpetra::Matrix<double,int,int,Node>& A,
-                                const Kokkos::View<bool*, typename Node::device_type> & dirichletRows,
-                                Kokkos::View<bool*, typename Node::device_type> dirichletCols,
-                                Kokkos::View<bool*, typename Node::device_type> dirichletDomain) {
-    MueLu::DetectDirichletColsAndDomains<double,int,int,Node>(A,dirichletRows,dirichletCols,dirichletDomain);
-  }
-
-
-  // Zeros out rows
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void
-  ZeroDirichletRows(RCP<Xpetra::Matrix<Scalar, LocalOrdinal, GlobalOrdinal, Node> >& A,
-                    const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows,
-                    Scalar replaceWith) {
-    using ATS        = Kokkos::ArithTraits<Scalar>;
-    using range_type = Kokkos::RangePolicy<LocalOrdinal, typename Node::execution_space>;
-
-    typename ATS::val_type impl_replaceWith = replaceWith;
-
-    auto localMatrix = A->getLocalMatrixDevice();
-    LocalOrdinal numRows = A->getLocalNumRows();
-
-    Kokkos::parallel_for("MueLu:Utils::ZeroDirichletRows", range_type(0,numRows),
-                         KOKKOS_LAMBDA(const LocalOrdinal row) {
-                           if (dirichletRows(row)) {
-                             auto rowView = localMatrix.row(row);
-                             auto length  = rowView.length;
-                             for (decltype(length) colID = 0; colID < length; colID++)
-                               rowView.value(colID) = impl_replaceWith;
-                           }
-                         });
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void
-  Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
-  ZeroDirichletRows(RCP<Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node> >& A,
-                    const Kokkos::View<const bool*, typename NO::device_type>& dirichletRows,
-                    Scalar replaceWith) {
-    MueLu::ZeroDirichletRows<Scalar,LocalOrdinal,GlobalOrdinal,Node>(A, dirichletRows, replaceWith);
-  }
-
-  template <class Node>
-  void
-  Utilities_kokkos<double,int,int,Node>::
-  ZeroDirichletRows(RCP<Xpetra::Matrix<double, int, int, Node> >& A,
-                    const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows,
-                    double replaceWith) {
-    return MueLu::ZeroDirichletRows<double,int,int,Node>(A, dirichletRows, replaceWith);
-  }
-
-
-  // Zeros out rows
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void
-  ZeroDirichletRows(RCP<Xpetra::MultiVector<Scalar, LocalOrdinal, GlobalOrdinal, Node> >& X,
-                    const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows,
-                    Scalar replaceWith) {
-    using ATS        = Kokkos::ArithTraits<Scalar>;
-    using range_type = Kokkos::RangePolicy<LocalOrdinal, typename Node::execution_space>;
-
-    typename ATS::val_type impl_replaceWith = replaceWith;
-
-    auto myCols = X->getDeviceLocalView(Xpetra::Access::ReadWrite);
-    size_t numVecs = X->getNumVectors();
-    Kokkos::parallel_for("MueLu:Utils::ZeroDirichletRows_MV", range_type(0,dirichletRows.size()),
-                         KOKKOS_LAMBDA(const size_t i) {
-                           if (dirichletRows(i)) {
-                             for(size_t j=0; j<numVecs; j++)
-                               myCols(i,j) = impl_replaceWith;
-                           }
-                         });
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void
-  Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
-  ZeroDirichletRows(RCP<Xpetra::MultiVector<Scalar, LocalOrdinal, GlobalOrdinal, Node> >& X,
-                    const Kokkos::View<const bool*, typename NO::device_type>& dirichletRows,
-                    Scalar replaceWith) {
-    MueLu::ZeroDirichletRows<Scalar,LocalOrdinal,GlobalOrdinal,Node>(X, dirichletRows, replaceWith);
-  }
-
-  template <class Node>
-  void
-  Utilities_kokkos<double,int,int,Node>::
-  ZeroDirichletRows(RCP<Xpetra::MultiVector<typename Teuchos::ScalarTraits<Scalar>::magnitudeType, int, int, Node> >& X,
-                    const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows,
-                    double replaceWith) {
-    return MueLu::ZeroDirichletRows<double,int,int,Node>(X, dirichletRows, replaceWith);
-  }
-
-
-  // Zeros out columns
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void
-  ZeroDirichletCols(RCP<Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node> >& A,
-                    const Kokkos::View<const bool*, typename Node::device_type>& dirichletCols,
-                    Scalar replaceWith) {
-    using ATS        = Kokkos::ArithTraits<Scalar>;
-    using range_type = Kokkos::RangePolicy<LocalOrdinal, typename Node::execution_space>;
-
-    typename ATS::val_type impl_replaceWith = replaceWith;
-
-    auto localMatrix = A->getLocalMatrixDevice();
-    LocalOrdinal numRows = A->getLocalNumRows();
-
-    Kokkos::parallel_for("MueLu:Utils::ZeroDirichletCols", range_type(0,numRows),
-                         KOKKOS_LAMBDA(const LocalOrdinal row) {
-                           auto rowView = localMatrix.row(row);
-                           auto length  = rowView.length;
-                           for (decltype(length) colID = 0; colID < length; colID++)
-                             if (dirichletCols(rowView.colidx(colID))) {
-                               rowView.value(colID) = impl_replaceWith;
-                             }
-                         });
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void
-  Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
-  ZeroDirichletCols(RCP<Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node> >& A,
-                    const Kokkos::View<const bool*, typename NO::device_type>& dirichletCols,
-                    Scalar replaceWith) {
-    MueLu::ZeroDirichletCols<Scalar,LocalOrdinal,GlobalOrdinal,Node>(A, dirichletCols, replaceWith);
-  }
-
-  template <class Node>
-  void
-  Utilities_kokkos<double,int,int,Node>::
-  ZeroDirichletCols(RCP<Xpetra::Matrix<double,int,int,Node> >& A,
-                    const Kokkos::View<const bool*, typename Node::device_type>& dirichletCols,
-                    double replaceWith) {
-    return MueLu::ZeroDirichletCols<double,int,int,Node>(A, dirichletCols, replaceWith);
-  }
-
-  // Applies rowsum criterion
-  template<class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void ApplyRowSumCriterion(const Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node>& A,
-                            const typename Teuchos::ScalarTraits<Scalar>::magnitudeType rowSumTol,
-                            Kokkos::View<bool*, typename Node::device_type> & dirichletRows)
-  {
-    typedef Teuchos::ScalarTraits<Scalar> STS;
-    RCP<const Xpetra::Map<LocalOrdinal,GlobalOrdinal,Node>> rowmap = A.getRowMap();
-
-    auto dirichletRowsHost = Kokkos::create_mirror_view(dirichletRows);
-    Kokkos::deep_copy(dirichletRowsHost, dirichletRows);
-
-    for (LocalOrdinal row = 0; row < Teuchos::as<LocalOrdinal>(rowmap->getLocalNumElements()); ++row) {
-      size_t nnz = A.getNumEntriesInLocalRow(row);
-      ArrayView<const LocalOrdinal> indices;
-      ArrayView<const Scalar> vals;
-      A.getLocalRowView(row, indices, vals);
-
-      Scalar rowsum = STS::zero();
-      Scalar diagval = STS::zero();
-      for (LocalOrdinal colID = 0; colID < Teuchos::as<LocalOrdinal>(nnz); colID++) {
-        LocalOrdinal col = indices[colID];
-        if (row == col)
-          diagval = vals[colID];
-        rowsum += vals[colID];
-      }
-      if (STS::real(rowsum) > STS::magnitude(diagval) * rowSumTol)
-        dirichletRowsHost(row) = true;
-    }
-
-    Kokkos::deep_copy(dirichletRows, dirichletRowsHost);
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void
-  Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
-  ApplyRowSumCriterion(const Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node>& A,
-                       const typename Teuchos::ScalarTraits<Scalar>::magnitudeType rowSumTol,
-                       Kokkos::View<bool*, typename Node::device_type> & dirichletRows)
-  {
-    MueLu::ApplyRowSumCriterion<Scalar, LocalOrdinal, GlobalOrdinal, Node>(A,rowSumTol,dirichletRows);
-  }
-
-
-  template <class Node>
-  void
-  Utilities_kokkos<double,int,int,Node>::
-  ApplyRowSumCriterion(const Xpetra::Matrix<double,int,int,Node>& A,
-                       const typename Teuchos::ScalarTraits<double>::magnitudeType rowSumTol,
-                       Kokkos::View<bool*, typename Node::device_type> & dirichletRows)
-  {
-    MueLu::ApplyRowSumCriterion<double, int, int, Node>(A,rowSumTol,dirichletRows);
-  }
-
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  Teuchos::RCP<Xpetra::Vector<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node> > ReverseCuthillMcKee(const Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node> &Op) {
-    using local_matrix_type = typename Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node>::local_matrix_type;
-    using local_graph_type  = typename local_matrix_type::staticcrsgraph_type;
-    using lno_nnz_view_t    = typename local_graph_type::entries_type::non_const_type;
-    using device            = typename local_graph_type::device_type;
-    using execution_space   = typename local_matrix_type::execution_space;
-    using ordinal_type      = typename local_matrix_type::ordinal_type;
-
-    local_graph_type localGraph = Op.getLocalMatrixDevice().graph;
-
-    lno_nnz_view_t rcmOrder = KokkosGraph::Experimental::graph_rcm
-      <device, typename local_graph_type::row_map_type, typename local_graph_type::entries_type, lno_nnz_view_t>
-      (localGraph.row_map, localGraph.entries);
-
-    RCP<Xpetra::Vector<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node> > retval =
-      Xpetra::VectorFactory<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node>::Build(Op.getRowMap());
-
-    // Copy out and reorder data
-    auto view1D = Kokkos::subview(retval->getDeviceLocalView(Xpetra::Access::ReadWrite),Kokkos::ALL (), 0);
-    Kokkos::parallel_for("Utilities_kokkos::ReverseCuthillMcKee",
-                         Kokkos::RangePolicy<ordinal_type, execution_space>(0, localGraph.numRows()),
-                         KOKKOS_LAMBDA(const ordinal_type rowIdx) {
-                           view1D(rcmOrder(rowIdx)) = rowIdx;
-                         });
-    return retval;
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  Teuchos::RCP<Xpetra::Vector<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node> > CuthillMcKee(const Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node> &Op) {
-    using local_matrix_type = typename Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node>::local_matrix_type;
-    using local_graph_type  = typename local_matrix_type::staticcrsgraph_type;
-    using lno_nnz_view_t    = typename local_graph_type::entries_type::non_const_type;
-    using device            = typename local_graph_type::device_type;
-    using execution_space   = typename local_matrix_type::execution_space;
-    using ordinal_type      = typename local_matrix_type::ordinal_type;
-
-    local_graph_type localGraph = Op.getLocalMatrixDevice().graph;
-    LocalOrdinal numRows = localGraph.numRows();
-
-    lno_nnz_view_t rcmOrder = KokkosGraph::Experimental::graph_rcm
-      <device, typename local_graph_type::row_map_type, typename local_graph_type::entries_type, lno_nnz_view_t>
-      (localGraph.row_map, localGraph.entries);
-
-    RCP<Xpetra::Vector<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node> > retval =
-      Xpetra::VectorFactory<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node>::Build(Op.getRowMap());
-
-    // Copy out data
-    auto view1D = Kokkos::subview(retval->getDeviceLocalView(Xpetra::Access::ReadWrite),Kokkos::ALL (), 0);
-    // Since KokkosKernels produced RCM, also reverse the order of the view to get CM
-    Kokkos::parallel_for("Utilities_kokkos::ReverseCuthillMcKee",
-                         Kokkos::RangePolicy<ordinal_type, execution_space>(0, numRows),
-                         KOKKOS_LAMBDA(const ordinal_type rowIdx) {
-                           view1D(rcmOrder(numRows - 1 - rowIdx)) = rowIdx;
-                         });
-    return retval;
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  Teuchos::RCP<Xpetra::Vector<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node> >
-  Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::ReverseCuthillMcKee(const Matrix &Op) {
-    return MueLu::ReverseCuthillMcKee<Scalar,LocalOrdinal,GlobalOrdinal,Node>(Op);
-  }
-
-  template <class Node>
-  Teuchos::RCP<Xpetra::Vector<int,int,int,Node> >
-  Utilities_kokkos<double,int,int,Node>::ReverseCuthillMcKee(const Matrix &Op) {
-    return MueLu::ReverseCuthillMcKee<double,int,int,Node>(Op);
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  Teuchos::RCP<Xpetra::Vector<LocalOrdinal,LocalOrdinal,GlobalOrdinal,Node> >
-  Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::CuthillMcKee(const Matrix &Op) {
-    return MueLu::CuthillMcKee<Scalar,LocalOrdinal,GlobalOrdinal,Node>(Op);
-  }
-
-  template <class Node>
-  Teuchos::RCP<Xpetra::Vector<int,int,int,Node> >
-  Utilities_kokkos<double,int,int,Node>::CuthillMcKee(const Matrix &Op) {
-    return MueLu::CuthillMcKee<double,int,int,Node>(Op);
-  }
-
-  // Applies Ones-and-Zeros to matrix rows
-  // Takes a Boolean array.
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void
-  ApplyOAZToMatrixRows(Teuchos::RCP<Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node> >& A,
-                       const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows) {
-    TEUCHOS_ASSERT(A->isFillComplete());
-    using ATS        = Kokkos::ArithTraits<Scalar>;
-    using impl_ATS = Kokkos::ArithTraits<typename ATS::val_type>;
-    using range_type = Kokkos::RangePolicy<LocalOrdinal, typename Node::execution_space>;
-
-    RCP<const Xpetra::Map<LocalOrdinal,GlobalOrdinal,Node> > domMap = A->getDomainMap();
-    RCP<const Xpetra::Map<LocalOrdinal,GlobalOrdinal,Node> > ranMap = A->getRangeMap();
-    RCP<const Xpetra::Map<LocalOrdinal,GlobalOrdinal,Node> > Rmap = A->getRowMap();
-    RCP<const Xpetra::Map<LocalOrdinal,GlobalOrdinal,Node> > Cmap = A->getColMap();
-
-    TEUCHOS_ASSERT(static_cast<size_t>(dirichletRows.size()) == Rmap->getLocalNumElements());
-
-    auto localMatrix = A->getLocalMatrixDevice();
-    auto localRmap = Rmap->getLocalMap();
-    auto localCmap = Cmap->getLocalMap();
-
-    Kokkos::parallel_for("MueLu::Utils::ApplyOAZ",range_type(0,dirichletRows.extent(0)),
-                         KOKKOS_LAMBDA(const LocalOrdinal row) {
-                           if (dirichletRows(row)){
-                             auto rowView = localMatrix.row(row);
-                             auto length = rowView.length;
-                             auto row_gid = localRmap.getGlobalElement(row);
-                             auto row_lid = localCmap.getLocalElement(row_gid);
-
-                             for (decltype(length) colID = 0; colID < length; colID++)
-                               if (rowView.colidx(colID) == row_lid)
-                                 rowView.value(colID) = impl_ATS::one();
-                               else
-                                 rowView.value(colID) = impl_ATS::zero();
-                           }
-                         });
-  }
-
-  template <class Scalar, class LocalOrdinal, class GlobalOrdinal, class Node>
-  void
-  Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::
-  ApplyOAZToMatrixRows(Teuchos::RCP<Xpetra::Matrix<Scalar,LocalOrdinal,GlobalOrdinal,Node> >& A,
-                       const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows) {
-    MueLu::ApplyOAZToMatrixRows<Scalar,LocalOrdinal,GlobalOrdinal,Node>(A, dirichletRows);
-  }
-
-  template <class Node>
-  void
-  Utilities_kokkos<double,int,int,Node>::
-  ApplyOAZToMatrixRows(Teuchos::RCP<Xpetra::Matrix<double,int,int,Node> >& A,
-                       const Kokkos::View<const bool*, typename Node::device_type>& dirichletRows) {
-    MueLu::ApplyOAZToMatrixRows<double,int,int,Node>(A, dirichletRows);
-  }
 
 } //namespace MueLu
