+++ packages/muelu/test/unit_tests_kokkos/Utilities_kokkos.cpp
@@ -95,7 +95,7 @@ namespace MueLuTests
 
     // CM Test
     {
-      auto ordering = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::CuthillMcKee(*A);
+      auto ordering = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>::CuthillMcKee(*A);
 
       TEST_EQUALITY(ordering->getLocalLength(), A->getLocalNumRows());
       TEST_EQUALITY(ordering->getGlobalLength(), A->getGlobalNumRows());
@@ -103,7 +103,7 @@ namespace MueLuTests
 
     // RCM Test
     {
-      auto ordering = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>::ReverseCuthillMcKee(*A);
+      auto ordering = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>::ReverseCuthillMcKee(*A);
 
       TEST_EQUALITY(ordering->getLocalLength(), A->getLocalNumRows());
       TEST_EQUALITY(ordering->getGlobalLength(), A->getGlobalNumRows());
@@ -117,7 +117,7 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using TST = Teuchos::ScalarTraits<Scalar>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
 
     auto A = MueLu_TestHelper_Factory::Build1DPoisson(100);
@@ -125,7 +125,7 @@ namespace MueLuTests
 
     CreateDirichletRow<Scalar, LocalOrdinal, GlobalOrdinal, Node>(A, localRowToZero);
 
-    auto drows = Utils_Kokkos::DetectDirichletRows(*A);
+    auto drows = Utils::DetectDirichletRows_kokkos(*A);
     auto drowsHost = Kokkos::create_mirror_view(drows);
     Kokkos::deep_copy(drowsHost, drows);
 
@@ -135,7 +135,7 @@ namespace MueLuTests
     CreateDirichletRow<Scalar, LocalOrdinal, GlobalOrdinal, Node>(A, localRowToZero, Teuchos::as<Scalar>(0.25));
 
     // row 5 should not be Dirichlet
-    drows = Utils_Kokkos::DetectDirichletRows(*A, TST::magnitude(0.24));
+    drows = Utils::DetectDirichletRows_kokkos(*A, TST::magnitude(0.24));
     drowsHost = Kokkos::create_mirror_view(drows);
     Kokkos::deep_copy(drowsHost, drows);
 
@@ -143,7 +143,7 @@ namespace MueLuTests
     TEST_EQUALITY(drowsHost(localRowToZero - 1), false);
 
     // row 5 should be Dirichlet
-    drows = Utils_Kokkos::DetectDirichletRows(*A, TST::magnitude(0.26), true);
+    drows = Utils::DetectDirichletRows_kokkos(*A, TST::magnitude(0.26), true);
     drowsHost = Kokkos::create_mirror_view(drows);
     Kokkos::deep_copy(drowsHost, drows);
 
@@ -158,7 +158,7 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using TST = Teuchos::ScalarTraits<Scalar>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
 
     auto A = MueLu_TestHelper_Factory::Build1DPoisson(100);
@@ -166,8 +166,8 @@ namespace MueLuTests
     LocalOrdinal localRowToZero = 5;
     CreateDirichletRow<Scalar, LocalOrdinal, GlobalOrdinal, Node>(A, localRowToZero);
 
-    auto drows = Utils_Kokkos::DetectDirichletRows(*A);
-    auto dcols = Utils_Kokkos::DetectDirichletCols(*A, drows);
+    auto drows = Utils::DetectDirichletRows_kokkos(*A);
+    auto dcols = Utils::DetectDirichletCols(*A, drows);
 
     auto dcolsHost = Kokkos::create_mirror_view(dcols);
     Kokkos::deep_copy(dcolsHost, dcols);
@@ -186,11 +186,11 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using TST = Teuchos::ScalarTraits<Scalar>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils_kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
 
     auto A = MueLu_TestHelper_Factory::Build1DPoisson(100);
-    auto diag = Utils_Kokkos::GetMatrixDiagonal(*A);
+    auto diag = Utils_kokkos::GetMatrixDiagonal(*A);
     auto diagView = diag->getHostLocalView(Xpetra::Access::ReadOnly);
 
     TEST_EQUALITY(diagView.extent(0), A->getLocalNumRows());
@@ -209,11 +209,11 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using TST = Teuchos::ScalarTraits<Scalar>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils_kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
 
     auto A = MueLu_TestHelper_Factory::Build1DPoisson(100);
-    auto diag = Utils_Kokkos::GetMatrixDiagonalInverse(*A);
+    auto diag = Utils_kokkos::GetMatrixDiagonalInverse(*A);
     auto diagView = diag->getHostLocalView(Xpetra::Access::ReadOnly);
 
     TEST_EQUALITY(diagView.extent(0), A->getLocalNumRows());
@@ -231,11 +231,11 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using TST = Teuchos::ScalarTraits<Scalar>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils_kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
 
     auto A = MueLu_TestHelper_Factory::Build1DPoisson(100);
-    auto diag = Utils_Kokkos::GetMatrixOverlappedDiagonal(*A);
+    auto diag = Utils_kokkos::GetMatrixOverlappedDiagonal(*A);
     auto diagView = diag->getHostLocalView(Xpetra::Access::ReadOnly);
 
     for (size_t idx = 0; idx < diagView.extent(0); ++idx)
@@ -251,7 +251,7 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using TST = Teuchos::ScalarTraits<Scalar>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using RangeType = Kokkos::RangePolicy<LocalOrdinal, typename Node::execution_space>;
 
@@ -265,7 +265,7 @@ namespace MueLuTests
 
     // Zero-ed out Vector
     {
-      Utils_Kokkos::FindNonZeros(vector->getDeviceLocalView(Xpetra::Access::ReadOnly), nonZeros);
+      Utils::FindNonZeros(vector->getDeviceLocalView(Xpetra::Access::ReadOnly), nonZeros);
 
       Kokkos::parallel_reduce(
           "", RangeType(0, nonZeros.extent(0)),
@@ -278,7 +278,7 @@ namespace MueLuTests
     {
       vector->putScalar(TST::one());
 
-      Utils_Kokkos::FindNonZeros(vector->getDeviceLocalView(Xpetra::Access::ReadOnly), nonZeros);
+      Utils::FindNonZeros(vector->getDeviceLocalView(Xpetra::Access::ReadOnly), nonZeros);
 
       Kokkos::parallel_reduce(
           "", RangeType(0, nonZeros.extent(0)),
@@ -296,7 +296,7 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using TST = Teuchos::ScalarTraits<Scalar>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
 
     auto A = MueLu_TestHelper_Factory::Build1DPoisson(100);
@@ -304,11 +304,11 @@ namespace MueLuTests
     LocalOrdinal localRowToZero = 5;
     CreateDirichletRow<Scalar, LocalOrdinal, GlobalOrdinal, Node>(A, localRowToZero);
 
-    auto drows = Utils_Kokkos::DetectDirichletRows(*A);
+    auto drows = Utils::DetectDirichletRows_kokkos(*A);
     Kokkos::View<bool *, typename Node::device_type> dirichletCols("dirichletCols", A->getColMap()->getLocalNumElements());
     Kokkos::View<bool *, typename Node::device_type> dirichletDomain("dirichletDomain", A->getDomainMap()->getLocalNumElements());
 
-    Utils_Kokkos::DetectDirichletColsAndDomains(*A, drows, dirichletCols, dirichletDomain);
+    Utils::DetectDirichletColsAndDomains(*A, drows, dirichletCols, dirichletDomain);
 
     auto dCol_host = Kokkos::create_mirror_view(dirichletCols);
     Kokkos::deep_copy(dCol_host, dirichletCols);
@@ -328,7 +328,7 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using TST = Teuchos::ScalarTraits<Scalar>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
 
     auto A = MueLu_TestHelper_Factory::Build1DPoisson(100);
@@ -336,13 +336,13 @@ namespace MueLuTests
     LocalOrdinal localRowToZero = 5;
     CreateDirichletRow<Scalar, LocalOrdinal, GlobalOrdinal, Node>(A, localRowToZero);
 
-    auto drows = Utils_Kokkos::DetectDirichletRows(*A);
-    auto dcols = Utils_Kokkos::DetectDirichletCols(*A, drows);
+    auto drows = Utils::DetectDirichletRows_kokkos(*A);
+    auto dcols = Utils::DetectDirichletCols(*A, drows);
 
     // Matrix version
     {
       const auto zeroVal = Scalar(0.35);
-      Utils_Kokkos::ZeroDirichletRows(A, dcols, zeroVal);
+      Utils::ZeroDirichletRows(A, dcols, zeroVal);
 
       Teuchos::ArrayView<const LocalOrdinal> indices;
       Teuchos::ArrayView<const Scalar> values;
@@ -360,7 +360,7 @@ namespace MueLuTests
       auto vector = Xpetra::MultiVectorFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>::Build(map, A->getGlobalNumCols());
 
       const auto zeroVal = Scalar(0.25);
-      Utils_Kokkos::ZeroDirichletRows(vector, dcols, zeroVal);
+      Utils::ZeroDirichletRows(vector, dcols, zeroVal);
 
       auto vecView = vector->getHostLocalView(Xpetra::Access::ReadOnly);
 
@@ -379,7 +379,7 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using TST = Teuchos::ScalarTraits<Scalar>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using RangeType = Kokkos::RangePolicy<LocalOrdinal, typename Node::execution_space>;
 
@@ -394,7 +394,7 @@ namespace MueLuTests
         });
 
     const auto zeroVal = Scalar(0.25);
-    Utils_Kokkos::ZeroDirichletCols(A, dCols, zeroVal);
+    Utils::ZeroDirichletCols(A, dCols, zeroVal);
 
     const auto localMatrix = A->getLocalMatrixHost();
     const auto numRows = A->getLocalNumRows();
@@ -419,7 +419,7 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using Magnitude = typename Teuchos::ScalarTraits<Scalar>::magnitudeType;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
 
     auto A = MueLu_TestHelper_Factory::Build1DPoisson(100);
@@ -427,7 +427,7 @@ namespace MueLuTests
     const auto numRows = A->getLocalNumRows();
     Kokkos::View<bool *, typename Node::device_type> dRows("", numRows);
 
-    Utils_Kokkos::ApplyRowSumCriterion(*A, Magnitude(1.0), dRows);
+    Utils::ApplyRowSumCriterion(*A, Magnitude(1.0), dRows);
     auto dRowsHost = Kokkos::create_mirror_view(dRows);
     Kokkos::deep_copy(dRowsHost, dRows);
 
@@ -436,7 +436,7 @@ namespace MueLuTests
       TEST_EQUALITY(dRowsHost(idx), false);
     }
 
-    Utils_Kokkos::ApplyRowSumCriterion(*A, Magnitude(-1.0), dRows);
+    Utils::ApplyRowSumCriterion(*A, Magnitude(-1.0), dRows);
     Kokkos::deep_copy(dRowsHost, dRows);
 
     for (size_t idx = 0; idx < dRows.extent(0); ++idx)
@@ -452,7 +452,7 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using TST = Teuchos::ScalarTraits<Scalar>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
 
     auto A = MueLu_TestHelper_Factory::Build1DPoisson(100);
@@ -460,7 +460,7 @@ namespace MueLuTests
 
     Teuchos::ArrayRCP<Scalar> arr(100, Scalar(2.0));
 
-    Utils_Kokkos::MyOldScaleMatrix(*A, arr, false);
+    Utils::MyOldScaleMatrix(*A, arr, false);
 
     const auto localMatrixScaled = A->getLocalMatrixHost();
     const auto localMatrixOriginal = B->getLocalMatrixHost();
@@ -478,7 +478,7 @@ namespace MueLuTests
     }
 
 #ifdef HAVE_TPETRA_INST_INT_LONG_LONG
-    TEST_THROW(Utils_Kokkos::MyOldScaleMatrix_Epetra(*A, arr, false, false), MueLu::Exceptions::RuntimeError);
+    TEST_THROW(Utils::MyOldScaleMatrix_Epetra(*A, arr, false, false), MueLu::Exceptions::RuntimeError);
 #endif
 
   } //MyOldScaleMatrix
@@ -491,7 +491,7 @@ namespace MueLuTests
 
     using TST = Teuchos::ScalarTraits<Scalar>;
     using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using RangeType = Kokkos::RangePolicy<LocalOrdinal, typename Node::execution_space>;
 
@@ -503,9 +503,9 @@ namespace MueLuTests
           dRowsIn(index) = index % 2;
         });
 
-    Utils_Kokkos::ApplyOAZToMatrixRows(A, dRowsIn);
+    Utils::ApplyOAZToMatrixRows(A, dRowsIn);
 
-    auto dRowsOut = Utils_Kokkos::DetectDirichletRows(*A);
+    auto dRowsOut = Utils::DetectDirichletRows_kokkos(*A);
     auto dRowsOutHost = Kokkos::create_mirror_view(dRowsOut);
     Kokkos::deep_copy(dRowsOutHost, dRowsOut);
 
@@ -523,7 +523,7 @@ namespace MueLuTests
 
     using TST = Teuchos::ScalarTraits<Scalar>;
     using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils_kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
 
     auto A = MueLu_TestHelper_Factory::Build1DPoisson(100);
@@ -591,7 +591,7 @@ namespace MueLuTests
     TEST_EQUALITY(crsMat->getGlobalNumRows(), op->getGlobalNumRows());
     TEST_EQUALITY(crsMat->getLocalNumRows(), op->getLocalNumRows());
 
-    auto transposeRes = Utils_Kokkos::Transpose(*A);
+    auto transposeRes = Utils::Transpose(*A);
     TEST_EQUALITY(transposeRes->getGlobalNumRows(), A->getGlobalNumRows());
     TEST_EQUALITY(transposeRes->getLocalNumRows(), A->getLocalNumRows());
 
@@ -613,7 +613,7 @@ namespace MueLuTests
       auto vec = Teuchos::RCP<Vector>(VectorFactory::Build(map));
       vec->putScalar(Scalar(5.0));
 
-      auto inv = Utils_Kokkos::GetInverse(vec);
+      auto inv = Utils::GetInverse(vec);
 
       auto invData = inv->getData(0);
 
@@ -638,7 +638,7 @@ namespace MueLuTests
     {
       auto powerRes = Utils::PowerMethod(*A);
 
-      auto inversDiag = Utils_Kokkos::GetMatrixDiagonalInverse(*A);
+      auto inversDiag = Utils_kokkos::GetMatrixDiagonalInverse(*A);
       auto powerRes2 = Utils::PowerMethod(*A, inversDiag);
 
       TEST_INEQUALITY(powerRes, Scalar(0.0));
@@ -655,32 +655,32 @@ namespace MueLuTests
     MUELU_TESTING_LIMIT_SCOPE(Scalar, GlobalOrdinal, Node);
 
     using Magnitude = typename Teuchos::ScalarTraits<Scalar>::magnitudeType;
-    using Utils_Kokkos = MueLu::Utilities_kokkos<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
+    using Utils = MueLu::Utilities<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
     using MueLu_TestHelper_Factory = MueLuTests::TestHelpers_kokkos::TestFactory<Scalar, LocalOrdinal, GlobalOrdinal, Node>;
 
     auto A = MueLu_TestHelper_Factory::Build1DPoisson(100);
 
     ParameterList params;
 
-    auto coords = Utils_Kokkos::ExtractCoordinatesFromParameterList(params);
+    auto coords = Utils::ExtractCoordinatesFromParameterList(params);
     TEST_EQUALITY(coords, Teuchos::null);
 
     auto map = A->getMap();
     auto mvCoords = Xpetra::MultiVectorFactory<Magnitude, LocalOrdinal, GlobalOrdinal, Node>::Build(map, 1);
     params.set("Coordinates", mvCoords);
 #ifdef HAVE_TPETRA_INST_INT_LONG_LONG
-    coords = Utils_Kokkos::ExtractCoordinatesFromParameterList(params);
+    coords = Utils::ExtractCoordinatesFromParameterList(params);
     TEST_INEQUALITY(coords, Teuchos::null);
 #else
     // When compiled with INST_INT_INT (Epetra) function will throw,
     // as it expects "Coordinates" param to be of EpetraMultiVector type
-    TEST_THROW(Utils_Kokkos::ExtractCoordinatesFromParameterList(params), MueLu::Exceptions::RuntimeError);
+    TEST_THROW(Utils::ExtractCoordinatesFromParameterList(params), MueLu::Exceptions::RuntimeError);
 #endif
 
     auto comm = Parameters::getDefaultComm();
-    TEST_NOTHROW(Utils_Kokkos::SetRandomSeed(*comm));
+    TEST_NOTHROW(Utils::SetRandomSeed(*comm));
 
-    TEST_NOTHROW(Utils_Kokkos::MakeFancy(*(out.getOStream())));
+    TEST_NOTHROW(Utils::MakeFancy(*(out.getOStream())));
   } //UtilsFunctions
 
 #define MUELU_ETI_GROUP(SC, LO, GO, NO)                                                                 \
